requirements.txt
kafka-python==2.0.2
requests==2.27.1
psycopg2-binary==2.9.3
configparser==5.2.0